+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Model derivative downloads now use the new direct-s3 logic.
+- Migrated to Authentication API v2
 
 ## [9.0.0] - 2022-08-19